// src/app/model/RoleType.ts
export enum RoleType {
  PSC1 = "PSC1",
  PSE1 = "PSE1",
  PSE2 = "PSE2",
  PSE1_CYCLISTE = "PSE1_CYCLISTE",
  PSE2_CYCLISTE = "PSE2_CYCLISTE",
  PSE1_CAVALIER = "PSE1_CAVALIER",
  PSE2_CAVALIER = "PSE2_CAVALIER",
  CI = "CI",
  CDP = "CDP",
  CHEF_SECTEUR = "CHEF_SECTEUR",
  CDPMGE = "CDPMGE",
  CI_BSPP = "CI_BSPP",
  CI_ALPHA = "CI_ALPHA",
  CH_VPSP = "CH_VPSP",
  CH_VL = "CH_VL",
  CH_QUAD = "CH_QUAD",
  INFIRMIER = "INFIRMIER",
  MEDECIN = "MEDECIN",
  DLUS = "DLUS",
  REGULATEUR = "REGULATEUR",
  REGULATEUR_TERRAIN = "REGULATEUR_TERRAIN",
  AIDE_REGULATEUR = "AIDE_REGULATEUR",
  LOGISTICIEN = "LOGISTICIEN",
  ONYX = "ONYX",
  PARTICIPANT = "PARTICIPANT",
  PARTICIPANT_EXT = "PARTICIPANT_EXT"
}